Extract classifyLoads and helpers to classify-and-persist.js
Moves filterWasteBalanceRecords, classifyLoads, and
fetchPeriodicReportsSafe out of validate.js to bring the file
under SonarCloud's 500-line limit (S104).
